server/requirements.txt
aiohttp==3.7.4.post0
fire==0.4.0
numpy==1.21.0
python-dateutil==2.8.2
dataclasses==0.6
dataclasses_json==0.5.6
pygame==2.0.3
peewee==3.14.8
freetype-py==2.2.0
matplotlib==3.5.1
humanhash3==0.0.6
orjson==3.6.7
yappi==1.3.3
pendulum==2.1.2
